implements a recursive type for SNF SNF is no longer a wrapper around FOF
