+++ razor-fol/src/transform/pnf.rs
@@ -1,183 +1,10 @@
 /*! Defines Prenex Normal Form (PNF) formulae and implements an algorithm for converting
-[`FOF`] to [`PNF`].
+a [`FOF`] to [`PNF`].
 
 [`FOF`]: crate::syntax::FOF
 */
 use super::{Substitution, TermBased, VariableRenaming};
-use crate::syntax::{formula::*, Term};
-use crate::syntax::{FOF, V};
-
-/// Defines the quantifier-free portion of [`PNF`].
-#[derive(Clone, Debug)]
-pub enum QuantifierFree {
-    /// Is logical top (⊤) or truth.
-    Top,
-
-    /// Is logical bottom (⟘) or falsehood.
-    Bottom,
-
-    /// Is an atomic formula, wrapping an [`Atom`].
-    Atom(Atom),
-
-    /// Is an atomic equality, wrapping an [`Equals`].
-    Equals(Equals),
-
-    /// Is the negation of a formula, wrapping a [`Not`].
-    Not(Box<Not<QuantifierFree>>),
-
-    /// Is a conjunction of two formulae, wrapping an [`And`].
-    And(Box<And<QuantifierFree>>),
-
-    /// Is a disjunction of two formulae, wrapping an [`Or`].
-    Or(Box<Or<QuantifierFree>>),
-
-    /// Is an implication between two formulae, wrapping an [`Implies`].
-    Implies(Box<Implies<QuantifierFree>>),
-
-    /// Is an bi-implication between two formulae, wrapping an [`Iff`].    
-    Iff(Box<Iff<QuantifierFree>>),
-}
-
-impl From<Atom> for QuantifierFree {
-    fn from(value: Atom) -> Self {
-        QuantifierFree::Atom(value)
-    }
-}
-
-impl From<Equals> for QuantifierFree {
-    fn from(value: Equals) -> Self {
-        QuantifierFree::Equals(value)
-    }
-}
-
-impl From<Not<QuantifierFree>> for QuantifierFree {
-    fn from(value: Not<QuantifierFree>) -> Self {
-        QuantifierFree::Not(Box::new(value))
-    }
-}
-
-impl From<And<QuantifierFree>> for QuantifierFree {
-    fn from(value: And<QuantifierFree>) -> Self {
-        QuantifierFree::And(Box::new(value))
-    }
-}
-
-impl From<Or<QuantifierFree>> for QuantifierFree {
-    fn from(value: Or<QuantifierFree>) -> Self {
-        QuantifierFree::Or(Box::new(value))
-    }
-}
-
-impl From<Implies<QuantifierFree>> for QuantifierFree {
-    fn from(value: Implies<QuantifierFree>) -> Self {
-        QuantifierFree::Implies(Box::new(value))
-    }
-}
-
-impl From<Iff<QuantifierFree>> for QuantifierFree {
-    fn from(value: Iff<QuantifierFree>) -> Self {
-        QuantifierFree::Iff(Box::new(value))
-    }
-}
-
-impl TermBased for QuantifierFree {
-    fn transform(&self, f: &impl Fn(&Term) -> Term) -> Self {
-        match self {
-            QuantifierFree::Top | QuantifierFree::Bottom => self.clone(),
-            QuantifierFree::Atom(this) => Atom {
-                predicate: this.predicate.clone(),
-                terms: this.terms.iter().map(f).collect(),
-            }
-            .into(),
-            QuantifierFree::Equals(this) => Equals {
-                left: f(&this.left),
-                right: f(&this.right),
-            }
-            .into(),
-            QuantifierFree::Not(this) => Not {
-                formula: this.formula.transform(f),
-            }
-            .into(),
-            QuantifierFree::And(this) => And {
-                left: this.left.transform(f),
-                right: this.right.transform(f),
-            }
-            .into(),
-            QuantifierFree::Or(this) => Or {
-                left: this.left.transform(f),
-                right: this.right.transform(f),
-            }
-            .into(),
-            QuantifierFree::Implies(this) => Implies {
-                premise: this.premise.transform(f),
-                consequence: this.consequence.transform(f),
-            }
-            .into(),
-            QuantifierFree::Iff(this) => Iff {
-                left: this.left.transform(f),
-                right: this.right.transform(f),
-            }
-            .into(),
-        }
-    }
-
-    fn rename_vars(&self, renaming: &impl VariableRenaming) -> Self {
-        // this does not rename bound variables of the formula
-        self.transform(&|t: &Term| t.rename_vars(renaming))
-    }
-
-    fn substitute(&self, substitution: &impl Substitution) -> Self {
-        self.transform(&|t: &Term| t.substitute(substitution))
-    }
-}
-
-impl Formula for QuantifierFree {
-    fn free_vars(&self) -> Vec<&V> {
-        match self {
-            Self::Top => Vec::new(),
-            Self::Bottom => Vec::new(),
-            Self::Atom(this) => this.free_vars(),
-            Self::Equals(this) => this.free_vars(),
-            Self::Not(this) => this.free_vars(),
-            Self::And(this) => this.free_vars(),
-            Self::Or(this) => this.free_vars(),
-            Self::Implies(this) => this.free_vars(),
-            Self::Iff(this) => this.free_vars(),
-        }
-    }
-}
-
-impl From<QuantifierFree> for FOF {
-    fn from(value: QuantifierFree) -> Self {
-        match value {
-            QuantifierFree::Top => Self::Top,
-            QuantifierFree::Bottom => Self::Bottom,
-            QuantifierFree::Atom(this) => Self::Atom(this),
-            QuantifierFree::Equals(this) => Self::Equals(this),
-            QuantifierFree::Not(this) => FOF::not(this.formula.into()),
-            QuantifierFree::And(this) => {
-                let left: FOF = this.left.into();
-                let right: FOF = this.right.into();
-                left.and(right)
-            }
-            QuantifierFree::Or(this) => {
-                let left: FOF = this.left.into();
-                let right: FOF = this.right.into();
-                left.or(right)
-            }
-            QuantifierFree::Implies(this) => {
-                let pre: FOF = this.premise.into();
-                let cons: FOF = this.consequence.into();
-                pre.implies(cons)
-            }
-            QuantifierFree::Iff(this) => {
-                let left: FOF = this.left.into();
-                let right: FOF = this.right.into();
-                left.iff(right)
-            }
-        }
-    }
-}
+use crate::syntax::{formula::*, Term, FOF, V};
 
 /// Represents a formula in Prenex Normal Form (PNF).
 ///
@@ -304,7 +131,6 @@ impl TermBased for PNF {
     }
 
     fn rename_vars(&self, renaming: &impl VariableRenaming) -> Self {
-        // this does not rename bound variables of the formula
         self.transform(&|t: &Term| t.rename_vars(renaming))
     }
